+++ libs/annotation/src/utils/getPointsUpdatedAnnotation.ts
@@ -1,5 +1,5 @@
 import polylabel from 'polylabel'
-import { LINE_TEXT_POSITION_SPACING } from '../const'
+import { LINE_TEXT_POSITION_SPACING, POINT_TEXT_POSITION_SPACING } from '../const'
 
 import { getLineLength } from './getLineLength'
 import { getTextPosition } from './getTextPosition'
@@ -73,6 +73,14 @@ export function getPointsUpdatedAnnotation({
       const labelPosition = polylabel([currentPoints], 1) as LabelPosition
 
       currentAnnotation.points = currentPoints as [Point, Point]
+      currentAnnotation.labelPosition = labelPosition
+      break
+    }
+
+    case 'point': {
+      const labelPosition = [currentPoint[0], currentPoint[1] + POINT_TEXT_POSITION_SPACING] as LabelPosition
+
+      currentAnnotation.point = currentPoint
       currentAnnotation.labelPosition = labelPosition
       break
     }