+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [7.1.0](https://github.com/lunit-io/insight-viewer/compare/@lunit/insight-viewer@7.0.1...@lunit/insight-viewer@7.1.0) (2023-10-16)
+
+
+### Features
+
+* **viewer:** add annotation types about point ([5b90936](https://github.com/lunit-io/insight-viewer/commit/5b90936d69202e52e8a5227c393192cd4dcdce39))
+* **viewer:** add point text position spacing and apply to point logic ([8d57c2c](https://github.com/lunit-io/insight-viewer/commit/8d57c2c70b6f1bc2073037f7bb101126c7b6bdad))
+* **viewer:** add PointViewer and Drawer components ([cf41f46](https://github.com/lunit-io/insight-viewer/commit/cf41f46ff2d2f4ea63e2b849b17b043715cf70e4))
+* **viewer:** apply Point Annotation logic to annotation CRUD functions ([4f4810a](https://github.com/lunit-io/insight-viewer/commit/4f4810ac5b17cf5e192bdd26d952a6fc34b67216))
+* **viewer:** apply PointViewer, Drawer to Annotation Viewer, Drawer ([47c205b](https://github.com/lunit-io/insight-viewer/commit/47c205b468675a7ad0dede50b1d3d1812d36bb4f))
+
+
+### Bug Fixes
+
+* **viewer:** apply addedTargetAnnotation when onChange running ([8f3244c](https://github.com/lunit-io/insight-viewer/commit/8f3244c9e574a0470bf3324d200d1982b46bffe4))
+* **viewer:** modify className of Point Drawer, Viewer text tag ([5d01cb8](https://github.com/lunit-io/insight-viewer/commit/5d01cb83d0a054b58a8bb1e4a1a3fc301130ff21))
+* **viewer:** modify transform of Point path ([3f9eca9](https://github.com/lunit-io/insight-viewer/commit/3f9eca9b94befd07dd5dab6dea1ea7642753ed11))
+
 ## [7.0.1](https://github.com/lunit-io/frontend-components/compare/@lunit/insight-viewer@7.0.0...@lunit/insight-viewer@7.0.1) (2023-06-26)
 
 ### Features